Adds contrast between bug types
The first input will show an error in the console; the latter won't show any error, nor will it actually fire any event to the server. The error is on `bindKeys`, so presumably the issue i re binding the event targets; the first one must just end up with no target and the latter a missing one, or rather probably the `document` might have default listeners?